Renamed files and tracked all files
