+++ news_app/src/App.css
@@ -69,7 +69,13 @@
 }
 .tool_bar input{
   background-color:rgba(231, 225, 197, 0.993) ;
-}
+  box-shadow: 0px 2px 2px rgba(231, 225, 197, 0.993);
+  margin: 10px 10px;
+  padding: 10px 2px;
+  border-radius: 5px;
+  border: none;
+  cursor: pointer;
+ }
 
 
 .profile{
@@ -116,10 +122,7 @@
   color: rgb(61, 105, 88);
   font-weight: bold;
   margin-left: 20%;
-}
-
-.login {
-  
+  cursor: pointer;
 }
